deps: sdk: adapt TypedEvent inference to ethers-v5@^8
@typechain/ethers-v5 v8 changed the way contracts events are represented
in files: instead of 2 generic parameters (for tuple/array and object
version), there's now a TypedEvent interface, which then has the tuple
and object generic parameters, and the events are explicitly typed in
the generated contracts output. This commit fixes inferring this
generic parameter, and also uses it in the FilteredLog->LogWithEvent
inner type, which is used to type the log output of the utils.