sdk: include validTill in PFS and warn when using invalid
Although the list of auto-fetched PFSs take into account validity of
their deposits and expire PFSs which become invalid, the user can
specify an invalid one on CLI/config. If they try to use one with
invalid/expired registration, they may have issues as Matrix server may
refuse to update them on users presences. If/when they try to use an
invalid PFS, we log a warning in the console.