dapp: fix version validation and comparison
The validation and comparison of version numbers for the update procedure was
broken. Reason was an updated version of the library that is used to work with
semantic version strings. In fact the interface of this library has changed (to
the better). For some reason the typing system did not detect this issue. The
fix is basically just an adaption of the imports from this library.