dapp: export version information into store module
The property of the root store that hold all information related to the version
was already a "complex" object. There was a specific sub-set of mutations and
getters just related to this property. This is a perfect indicator that this can
be exported as its own independent store module. This change got triggered by
the knowledge that these version information will become extended soon.
Therefore this commit makes the refactoring to allow for better coding of the
upcoming feature.