dapp: rework service worker message encoding
The service worker and the service worker assistant are exchanging messages.
Therefore it is necessary to define a shared message en- and decoding. Messages
include a fixed string (using an enumeration to match them) as identifier which
type of message this is. Furthermore than can attach any kind of more payload to
it. Unfortunately the last part has been implemented incorrectly in the past.
This has not been detected yet because the payload was never evaluated so far.
The only message that uses a payload so far is for installation errors by the
service worker. The attached error object has been ignored by the assistant so far.
Now it is necessary to have more messages with a payload where the assistant
actually needs to access this data. The former implementation was based on
a misunderstanding of the W3C Client interface to send messages. Instead of
sending the payload with the message so the receiver can access it, the payload
has been only transferred in ownership. But the receiver was not able to read
this data. Therefore it is necessary to encode all the information into the
`data` field of the message. So far this field was used for the fixed string to
identify the message type. To solve the issue the data field has become an
object with a required `messageIdentifier` property. This property acts the same
as the whole field before and is used to match messages. The rest of the data
object is then the optional payload that gets attached.