dapp: enable testing for store module persistence
Some Vuex store modules have a plugin that ensure that the modules state gets
persisted to a storage backend. This ensures that the state survives a reload of
the application. This feature is mostly implemented by a 3rd party library and
was never tested. But it is an important feature of the dApp, especially for
upcoming new store modules. Therefore we need to ensure that it works as intended.
To enable testing, the store module plugins get now created by a function. This
allows to specify an optional parameter to define the storage backend. Per
default this is `window.localStorage` as always. The tests for the store module
can now simply provide a mocked version of the storage backend. It is a clean
separation of modules with clear interfaces. No need to mock global variables in
an awkward manner. The test cases do not hide anything and can be read and
understood easily.