dapp: add information of installed version
Knowing which version of the dApp has been installed is an important piece of
information. Having this data allows for future features to work with different
versions and ensure strict manual updates.
To do so, the service worker is sending a new message to the service worker
assistant when it installed a new version and got activated. The message
includes the installed version. The assistant will forward this data and put it
into the version information store of the dApp.
Having this data allows to compare the available version on the server with the
actual locally installed version to determine if there is an update available.
Before this was compared to the loaded version, which must not be correct all
the time.
The installed version information in the store is persisted to the local storage
of the browser, as it gets only set when a new version gets installed. This
property is the only one that gets stored from the version information store
module. Everything else is and must be temporary.