dapp: rename update snackbar component
The simple reason for this is that this component will be responsible for
a wider set of just plain updates. It kind of already is. To reflect this better
in the naming, the component got renamed to a more generic name while still
maintaining its responsibility and relationship. This is a single commit to keep
the history cleaner and better to understand.