dapp: add detection when wrong version got loaded
Combining the now available information of the installed version and the active
version allows to determine if the correct version got loaded. If this is not
the case, the dApp gets blocked for the user to prevent him from using this
wrong version (he did not confirm to update to). Further measurements will be
added to resolve this situation.