dapp: add feature to reload to installed version
The dApp gets blocked for the user when the wrong version got loaded. It is
necessary to resolve this situation automatically. The approach to do so is to
unregister all service workers and reload the tab. This solution has been
derived from practical tests with force-reloads to bypass the service worker.
Once the dApp got loaded, the service worker assistant gets triggered to verify
if the correct version got loaded and if not he resolves it.

To allow for proper testing of this feature, it is necessary to mock all parts
of the global browser environment the service worker assistant is using. Such
include for example the service worker container or the fetch function. In the
past this has been tested with "complicated" global mocks in before and after
test suite hooks. This isn't a too nice approach and should be avoided where
possible. Therefore the assistant can be instantiated with an optional last
parameter to define any sub-set of environment objects. Tests do now precisely
define what they need individually and locally.