dapp: add strong message handling
The service worker thread and the service worker assistant in the client thread
exchange messages. Recently it was necessary to extend the message encoding to
add some payload to them. Before that change, messages where pure Strings. The
whole string has been used to identify which type of message this is to call the
respective handler. Now the message type is encoded into an Object with possible
additional payload.
This was a breaking change in how messages are formatted. With this change the
senders and receivers have been adapted to work with this new format. But what
has been forgotten was that the service worker version and the client version
can differ. In fact it is possible that a client of version X gets served by
a service worker of version Y where Y > X. Each time a new version of both gets
deployed to the public hosted server, the browser will load the new available
script of the service worker. But the locally installed client (cached assets
served by service worker) remains the old one. In result does the client send
messages to the service worker it does not understand and vice versa. Despite
the offline functionality this breaks all other features, including the update
mechanism.
To solve this problem it is necessary to establish a message formatting that is
version agnostic. The topic of having different versions of the same
event/message is typical and quite complex one. Especially in environments where
different units of the communication run on different versions. The here chosen
approach is to use a wrapper around the message and an internal mapping between
the week schema (JSON) and the stronger internal typing schema. This approach
requires to never rename any message property and only add new ones. Furthermore
must the set of required properties be fixed and does not change with new
versions. Only this way can an old receiver understand the message of a new
sender. But also new receivers must be able to handle messages by old senders
which is guaranteed due to the required fields.
Having this methodology in place allows to have no strict versioning of the
message scheme and allows for independent sender and receiver versions.
Unfortunately there is already a version out there which acts on the old
messages. The initial definition with the required fields etc are not known to
them. To work with these versions, it is necessary add an exception to the whole
messaging process, combining it with other approaches of message versioning.
The service worker is never older than the client. This is a premise. To work
with old clients before the new messaging format it must do two things. First it
must be able to parse both message formats to act on messages from old and new
clients. Second it must send messages in old and new format too. This means each
message the service worker sends get send twice to each client, once in each
format. Luckily this is only necessary for all message type which are supported
until now. All future upcoming message types will not be handled by old clients
anyway. This saves traffic in future and reduces the amount of necessary tests.
The only thing the client has to do is to ignore messages in the old format. As
the service workers sends these messages twice, it should not act on both of
them. Old clients which are already released will in turn only act on the
messages in the old format and ignore messages in the new format already
(luckily).
Having this mechanism once in place allows to establish the new enriched
messaging format important for new features. And it is open for any new
extensions of the messages, using the same format and being backwards compatible
to old sender and receiver.

Please note that the unit tests got adapted to don't use the message type
enumeration and avoid to use any messaging helper where possible. This should
ensure that the tests are very strict and do not get affected by code
refactoring tools. This is necessary because old client/service worker versions
will not get this refactoring and new versions would break compatibility.