dapp: extend loading wrong version protection
The feature to verify if the correct version got loaded was so far only working
for new versions. Loading a new version on top of an old one (i.e. <= `2.0.1`)
will not be detected. But this is actually possible as a new version gets loaded
that include function functionality to do so. The only issue so far was that the
mechanism relies on the information that the client persists the information
which version got loaded. Old version have not done this. But this information
can be used. Should mean when there is no installed version number available,
this means the installed version must be <= 2.0.1. If the active version is
newer than that, it is wrong.
This change required to adapt the update mechanism. So far was the installed
version number deleted during an update as the check happens faster than the
installation. As this would now lead to wrong implications, the installed
version number never gets deleted. Rather a new flag of the store gets persisted
that gets set during an update. When this flag is set, the check always passes
as there can't be a wrong version loaded. As soon as the installation of the
updated completed this flag gets reset again, allowing to detect wrong versions
on the next load again.