dapp: improve UI while update in progress
When an update is in progress, the new UI gets blocked for the user and
a progress spinner was shown. This disappears after the page reloads. But here
the actual installation actually happens. Having the new persisted information
that an update is in progress this can be lengthened over the reload. The UI
remains remains blocked until the service worker is done with the installation.
The snackbar with the spinner is also active during this time.
Furthermore this fixes the issue that after the latest changes the update is
available message was shown while the update was still being installed.