dapp: fix loading for the first time
The dApp has been enhanced with functionality to detect if a wrong version got
loaded that is different to the installed one. Within this feature is was
necessary to handle older versions which have not been prepared for such
a feature. Newer versions persist the information which version they have
installed. If this information is not present, the dApp thinks it must be an old
installation. If it then is itself a client of the newer generation, it detects
itself as a wrong version. The problem is that this happens each time a newer
version gets loaded by a user who never used an old version before. There is
also no installed version information yet, but the loaded version is a new one.
The implication that this is wrong is a mistake.
To resolve this issue it is necessary to determine if the dApp has been loaded
for the very first time, no matter anything else. This is like the state of an
update in progress, it can't be the wrong version then.
The challenge is to determine if the dApp was loaded for the first time.
Especially because it must work also with the old versions again we can't touch.
The final solution detects if the user has ever accepted the disclaimer. This
isn't 100% the same as loading, but close enough and the best we could get.

This also includes a tiny change to call the mechanism to detect wrongly loaded
versions as soon as possible. The user must not be able to accept the disclaimer
before that.