dapp: remove all occurrences of settle timeout
In the past, channels had a settle timeout property. It was also possible to
configure the SDK with a settle timeout value for new channels. This has changed
with the latest version of the contracts and SDK. The settle timeout has moved
from individual channels to the token network itself. In result channel data
structure have no more this data field and the SDK doesn't accept this
configuration option anymore.
There is also a visual effect of this change. So far, if a channel got closed,
the button to settle the channel was showing a countdown until it becomes
possible. But not only was the settle timeout moved, it also changed its meaning
from a block number based value to an actual time value in seconds. The issue is
that this value can very from a few seconds to multiple days. In result is is
not easy to maintain a countdown on the settle button. Also because the
rendering of the Vue component would need to be force updated by an interval as
it is now purely time based. To maintain the raw functionality for the moment,
the button just remains disabled until the channel is settlable. This might
improve in future again.