dapp: adapt to new UDC withdraw plans
In the past, a planned UDC withdrawal had the property of a target `withdrawBlock`
from which it is possible to withdraw the tokens. This has changed to a timestamp
value (in Unix format) called `withdrawableAfter`. Due to this change, a couple
of renaming where necessary.
Furthermore this also has visual effects. On the view to the user deposit
contract, if there was a planned withdraw pending (not yet withdrawable) it was
showing a countdown for the blocks remaining. As this is not the case anymore,
it has been changed to display the target date. Once this date has passed, the
UI changes as usual and displays the user that the plan is now ready.

There was also a tiny change in the translations, not only according block to
time change. Also a special unicode character got replaced with a simple ASCII
one, as the former one was not rendered properly for some users, depending on
their system font.