replace python with light client for e2e tests
The Python Client becomes discontinued. Thereby it will not support future
contract versions or any other changes to the protocol. To maintain a future
proof end-to-end environment, it is necessary to replace the partner Raiden
nodes with Light Client CLI nodes.

The nature of this commit is very chaotic. This is due to the fact that the
whole end-to-end environment setup is not very well made and breaking changes
like this must result into a lot of issues (as we had). As compromise the made
changes do not try to improve the overall setup much. There are a few tiny
improvements in regards of Docker build caching. A lot of work went into the
untangling of the Python virtual environments. The removal of the Python Client
environment cause issue due to hidden dependencies and the overall virtual
environment management.

Within this commit, the `EXPIRY_FACTOR` got increased for the dApp in `e2e`
mode. The defaults result into a very short reveal timeout margins that can lead
to flakiness on slower developer computers. Increasing it makes the tests more
save to run through successfully.