update e2e environment for new timestamp contracts
This means that basically all related components like contracts, services etc.
got updated. As the contracts are not released yet and thereby all dependent
components, all versions point to pre-release git tags.
There were a few adaption necessary despite exchanging the components which come
with changes of these updates.

As a side-effect we get a eat-your-own-dogs-food dilemma here. As there is no
Raiden client that supports the new contracts yet, we need to fetch the current
under-development branch of the Light Client. This must change once a release is
ready.

The end-to-end tests of the SDK and dApp do not succeed yet. They have not been
touched and are likely to fail by syntax and semantic issues.