sdk: add matrixServer to PFS info and use it in auto-matrixServer
When matrixServer isn't set, we are in `auto` mode. Previously, we'd
fetch the list from `matrixServerLookup`, but is easy to have this list
mismatched with current network/deployment. This change prefers the same
matrixServer from the PFSs, scanning first `config.additionalServices`
and then services fetched from ServicesRegistry, which are expected to
always be compatible with current deployment.