fix too quick channel interactions for quick-pay
When using the quick-pay feature and being connected to an Arbitrum
based network, this can lead to issues when a new channel must be opened
or a deposit to an existing one. In both cases there is an initial
blockchain interaction, followed by a transfer via the just "modified"
channel. For a successful payment this requires a fully synced channel
partner. Due to the zero confirmation blocks on Arbitrum, this is not
always the case. Especially if the payment message gets send very quick
after the blockhain interaction. The partner node sometimes does not
know about the new channel or deposit yet and will deny/errors (on) the
sent message. In result the whole quick-pay will fail. This is not an
issue for regular dApp usage where user manually opens/deposits (to)
a channel and then triggers a payment.
This issue did already occur in the scenario player nightly tests. The
well working solution approach is a short delay after channel blockchain
interactions. Thereby this get applied here for the dApp too. The
additional delay doesn't add much in case of a non Arbitrum network with
confirmation blocks. Therefore no further conditional optimization is
necessary here.